perf: enable zero-copy KifDataset tensors

+++ tests/maou/app/learning/test_dataset.py
@@ -30,8 +30,8 @@ def test_preprocessed_batches_provide_legal_move_masks() -> None:
         [
             ("boardIdPositions", np.uint8, (9, 9)),
             ("piecesInHand", np.uint8, (14,)),
-            ("moveLabel", np.float16, (5,)),
-            ("resultValue", np.float16),
+            ("moveLabel", np.float32, (5,)),
+            ("resultValue", np.float32),
         ]
     )
     data = np.array(
@@ -39,14 +39,14 @@ def test_preprocessed_batches_provide_legal_move_masks() -> None:
             (
                 np.eye(9, dtype=np.uint8),
                 np.zeros(14, dtype=np.uint8),
-                np.array([1, 0, 0, 0, 0], dtype=np.float16),
-                np.float16(1.0),
+                np.array([1, 0, 0, 0, 0], dtype=np.float32),
+                np.float32(1.0),
             ),
             (
                 np.fliplr(np.eye(9, dtype=np.uint8)),
                 np.zeros(14, dtype=np.uint8),
-                np.array([0, 1, 0, 0, 0], dtype=np.float16),
-                np.float16(-1.0),
+                np.array([0, 1, 0, 0, 0], dtype=np.float32),
+                np.float32(-1.0),
             ),
         ],
         dtype=dtype,
@@ -58,10 +58,10 @@ def test_preprocessed_batches_provide_legal_move_masks() -> None:
     (boards, pieces), (_, _, legal_move_mask) = next(iter(loader))
 
     assert isinstance(boards, torch.Tensor)
-    assert boards.dtype == torch.long
+    assert boards.dtype == torch.uint8
     assert boards.shape == (2, 9, 9)
     assert isinstance(pieces, torch.Tensor)
-    assert pieces.dtype == torch.float32
+    assert pieces.dtype == torch.uint8
     assert pieces.shape == (2, 14)
     assert isinstance(legal_move_mask, torch.Tensor)
     assert torch.all(legal_move_mask == 1)
